Removed maintenance module. Implemented login check for siteadmin. Siteadmin paging bug fixes.

+++ apps/courses/modules/admincourse/templates/_list.php
@@ -9,14 +9,20 @@
     <?php foreach ($course_list->getResults() as $course): ?>
     <tr>
       <td width="30">
-      <?php //FIXME selecting an instructor will cause paging to reorder?>
-      <?php echo link_to(" ", 'admincourse/delete?id='.$course->getId(), array('method' => 'delete', 'class'=>'deletebtn', 'confirm' => 'Are you sure?')) ?>
-      <?php echo link_to(" ", 'admincourse/edit?id='.$course->getId(), array('class'=>'select'))?>
+      <?php if($sf_request->getParameter('page') ===null):?>
+	      <?php echo link_to(" ", 'admincourse/delete?id='.$course->getId(), array('method' => 'delete', 'class'=>'deletebtn', 'confirm' => 'Are you sure?')) ?>
+	      <?php echo link_to(" ", 'admincourse/edit?id='.$course->getId(), array('class'=>'select'))?>
+	  <?php else:?>
+	      <?php echo link_to(" ", 'admincourse/delete?id='.$course->getId().'&page='.$sf_request->getParameter('page'), array('method' => 'delete', 'class'=>'deletebtn', 'confirm' => 'Are you sure?')) ?>
+	      <?php echo link_to(" ", 'admincourse/edit?id='.$course->getId().'&page='.$sf_request->getParameter('page'), array('class'=>'select'))?>
+      <?php endif;?>
       </td>
-      <td><?php echo $course->getId() ?></td>
+      <td<?php if ($sf_request->hasParameter('id') && $sf_request->getParameter('id')==$course->getId()):?> style='background:#FFE87C'<?php endif;?>>
+        <?php echo $course->getId() ?>
+      </td>
     </tr>
     <?php endforeach; ?>
     
-    <?php include_partial('global/paging', array('pagelist' => $course_list, 'location' => 'admincourse')) ?>
+    <?php include_partial('global/paging', array('pagelist' => $course_list, 'location' => skuleadminConst::decomposeURL($sf_context, $sf_request))) ?>
   </tbody>
 </table>
