+++ apps/courses/modules/admininstructor/templates/_list.php
@@ -9,15 +9,19 @@
     <?php foreach ($instructor_list->getResults() as $instructor): ?>
     <tr>
 	    <td width="30">
-	    <?php //FIXME selecting an instructor will cause paging to reorder?>
+	    <?php if($sf_request->getParameter('page') ===null):?>
 	    	<?php echo link_to(" ", 'admininstructor/delete?id='.$instructor->getId(), array('method' => 'delete', 'class'=>'deletebtn', 'confirm' => 'Are you sure?')) ?>
             <?php echo link_to(" ", 'admininstructor/edit?id='.$instructor->getId(), array('class'=>'select'))?>
+        <?php else:?>
+        	<?php echo link_to(" ", 'admininstructor/delete?id='.$instructor->getId().'&page='.$sf_request->getParameter('page'), array('method' => 'delete', 'class'=>'deletebtn', 'confirm' => 'Are you sure?')) ?>
+            <?php echo link_to(" ", 'admininstructor/edit?id='.$instructor->getId().'&page='.$sf_request->getParameter('page'), array('class'=>'select'))?>
+        <?php endif;?>
 	    </td>
-	    <td>
+	    <td<?php if ($sf_request->hasParameter('id') && $sf_request->getParameter('id')==$instructor->getId()):?> style='background:#FFE87C'<?php endif;?>>
 	    	<?php echo $instructor->getLastName() ?>, <?php echo $instructor->getFirstName() ?>
 	    </td>
     </tr>
     <?php endforeach; ?>
-    <?php include_partial('global/paging', array('pagelist' => $instructor_list, 'location' => 'admininstructor')) ?>
+    <?php include_partial('global/paging', array('pagelist' => $instructor_list, 'location' => skuleadminConst::decomposeURL($sf_context, $sf_request))) ?>
   </tbody>
 </table>
